+++ packages/Button/src/components/Close/Close.js
@@ -2,7 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import CloseIcon from "@paprika/icon/lib/Times";
 import * as constants from "@paprika/constants/lib/Constants";
-import * as types from "./types";
+import * as types from "../../types";
 import * as sc from "./CloseButton.styles";
 
 const propTypes = {
@@ -18,7 +18,7 @@ const defaultProps = {
   isDark: false,
 };
 
-const CloseButton = React.forwardRef((props, ref) => {
+const Close = React.forwardRef((props, ref) => {
   const buttonProps = {
     a11yText: props.a11yText || "close", // TODO: use L10n
     isDark: props.isDark,
@@ -32,11 +32,11 @@ const CloseButton = React.forwardRef((props, ref) => {
   );
 });
 
-CloseButton.types = {
+Close.types = {
   kind: constants.kind,
 };
-CloseButton.displayName = "CloseButton";
-CloseButton.propTypes = propTypes;
-CloseButton.defaultProps = defaultProps;
+Close.displayName = "Button.Close";
+Close.propTypes = propTypes;
+Close.defaultProps = defaultProps;
 
-export default CloseButton;
+export default Close;