+++ packages/Button/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+Move CloseButton, IconButton, LinkButton to components folder to fix typings and for file structure consistency [kcumlat](https://github.com/kcumlat).
+
 ## [0.3.0] - 2020-06-16
 
 ### Changed
@@ -25,3 +27,7 @@ Added support for icon in LinkButton component [@tristanjasper](https://github.c
 ## [0.3.11] - 2020-09-17
 
 Added aria-disabled tag on LinkButton component [@tristanjasper](https://github.com/tristanjasper).
+
+## [0.3.12] - 2020-09-23
+
+Updated DEFAULT hover,visited styles to be resilient to global styling, [@tristanjasper](https://github.com/tristanjasper).